:zap: Speed up `lamin list branch` (#228)
