+++ .github/workflows/build.yml
@@ -97,4 +97,4 @@ jobs:
       - run: uv pip install --system git+https://github.com/laminlabs/laminprofiler
       - run: lamin login
       - run: lamin connect laminlabs/lamindata
-      - run: laminprofiler check tests/profiling/lamin_list_branch_django.py --threshold 2
+      - run: laminprofiler check tests/profiling/lamin_list_branch.py --threshold 1.55
