+++ tests/profiling/lamin_list_branch.py
@@ -1,0 +1,3 @@
+from lamin_cli.hub import list_branches
+
+print(list_branches())
