--- tests/profiling/lamin_list_branch_django.py
@@ -1,3 +1,0 @@
-from lamindb.models import Branch
-
-print(Branch.to_dataframe())
